Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The common case is two APNs in the db for this mcc/mnc.  One is the default and one is for mms.  They almost always have a different APN names.  If the code today is getting the APN name of the current connection (default) and passing that in to the query, then this change seems like it would cause mms to fail as we'd only match the default which doesn't have mmsc settings.  It seems like an bug to me that the mms code passes in the apn name and it was masked by the fact we tossed out the selection.  MMS could *could* use the apn name to prefer the current APN provided it has an mmsc value, but that's not what it does today.

I'm not opposed to this change - I think ignoring data passed into the query seems wrong.  However you'd need to fix call sites that are doing the wrong thing but getting away with it due to the existing code.

Patch-set: 1
Reviewer: Robert Greenwalt <1002609@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0