Update patch set 1
Patch Set 1:

TransactionService.java listens EVENT_DATA_STATE_CHANGED by NetworkConnectivityListener.
It checks whether network type is MMS or not when it gets event.
If the connected network type is MMS, it gets the apn name by calling getExtraInfo of NetworkInfo.
So there seems to be no chance that MMS passes defaul apn name to the query.

Patch-set: 1