Group contacts are displayed as anonymous
Root Cause:
The recipient ids of a group threads were saved as a string and
using SPACE to separate the ids in the string. When updating
canonical address if a thread is deleted, the address of a group
sms in the canonical_address will be deleted since it can not
match any recipients in the threads table.

Solution:
Compose a string that contains all recipient ids in the thread table,
and use ',' to separate the ids. Then, use this string in sql
statement to clear the canonical_address table.

Change-Id: I947a8b5a07eb98fce11e57eee4ea5c3772bc9f29