Update patch set 1
Patch Set 1: Fails

The underlying code has changed in Google's internal master tree, in a way that appears to handle this issue. I recommend to abandon this patch and upload a new one in the future if the issue still exists in the next major release.

As a note, the commit message should be wrapped to about 70 columns.

Patch-set: 1
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=-1