Update patch set 1
Patch Set 1:

Thanks for the fix! However, this bug was fixed in a similar manner on Dec. 19, 2011 (but not quite as elegantly as your fix): Deleting a thread will mess up a multi-recipient thread with the same recipient. The previous DB query for deleting unreferenced recipient id's didn't work on threads addressed to multiple recipients because the recipient id's are stored as a space-separated list of ids in a single column. With this change, the work for identifying unreferenced recipient id's is done in java, rather than in an sqlite3 selection statement. If the change isn't in master right now, it will be in the next push.

Patch-set: 1