Fixing problem with retrieving a single SMS from an ICC slot
The parameter to content://sms/icc/# is supposed to be
the one-based ICC record index, but instead it was
treated as the index of the local list holding the messages.
If the SIM has any empty or non-supported records,
these two indices will not match.

For example if the SIM has 20 slots of which only slot 7 and 11 are
used, SmsManager.getAllMessagesFromIcc will return a list of 2 elements.
When retrieving message 7, we must not try to get element number 7
in that list but rather iterate through the list to find the message
that has the _SIM record index_ 7.

Change-Id: I086e81d564f0eb1f65bd8f59abfa9ccab8225da0