Fix apn restore issue for multi-sim.
If "Restore to Default" option is selected on apn menu, then
all carriers table entries are deleted and apns file is read
again. Due to this, user added apns on both subs are lost though
restore is selected for one sub.
Fix is to delete all apns except for user added apns on the
other available sub.

Change-Id: I4c01e87829c7deca2cf29b479149d79d7e78c9c7