Update patch set 1
Patch Set 1:

I found this while checking an old contribution:
 https://android-review.googlesource.com/#/c/47321

I do not think this method is used very often, it seems to be broken the last couple of years. Are you still interested in the corrections in this method?

Patch-set: 1
Reviewer: Henrik Baard <1004424@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0