Fix apn overwrite issue
When a new apn is created with an existing apn name but
different apn type, existing apn is overwritten with new apn.
This is because of the unique constraint defined for carriers
table.
Fix is to add type field also to the unique constraint,
which would allow multiple apns to share same apn with different types.

Test: Manual - Checked that new apn added without overwriting
existing apn, when adding an apn that differs only by type field.
Bug: 36378643
Change-Id: I7d7c5aa6bea42cde9db8bb841802f8cd7e5a7095