Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

We're trying to maintain oversight and control of Telephony projects.  Tom, is it ok to bump you and Tim?  Messaging at this point should just be a client and not really need unfettered access to this project.

Note that the mms project doesn't seem to have you on it - that seems more useful than this provider.

Patch-set: 3
Reviewer: Robert Greenwalt <1002609@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1