Fix apn restore issue for multi-sim
If "Reset to default" option is selected on apn menu, then
all carriers table entries are deleted and apns file is read
again. Due to this, user added apns on both subs are lost though
restore is selected for one sub.
Fix is to delete only the edited apns except sub.

Test: manual - Checked that other SIM's apn does not lost after
executing "Reset to default"
Bug: 65188023

Change-Id: I4c01e87829c7deca2cf29b479149d79d7e78c9c7