Fix apn restore issue for multi-sim
If "Reset to default" option is selected on apn menu, then
all carriers table entries are deleted and apns file is read
again. Due to this, user added apns on both subs are lost though
restore is selected for one sub.
Fix is to delete apns except other sub.

Test: manual - Checked that other SIM's apn does not lost after
executing "Reset to default"
Test: auto - Passed TelephonyProviderTest
Bug: 65188023

Change-Id: Idbccd47c79c0197ddb07ab1db930e6058b111c0f