Add support for updating apns when only resource changed
Apn added in RRO is not added when rebooting after firmware upgrade.

On initial boot the apn database will consist of a merge between:
com.android.internal.R.xml.apns and one of the files below;
- /system/etc/apns-conf.xml
- /oem/telephony/apns-conf.xml
- /data/misc/apns-conf.xml

For the database to be upgraded after a firmware upgrade, one of the three
apns-conf files above need to be changed, no consideration is given
that the RRO might have been updated.

This commit adds the checksum for resources to the selected file for
a total checksum that is then compared to persisted value.

Bug: 79876418

Test: Start device and check the apns in settings
Add an RRO that contains another set of apns
Remove /data/user_de/0/com.android.providers.telephony/shared_prefs/build-id.xml
to simulate that the fingerprint is updated
Restart the device and the new matching apns exists in settings

Change-Id: I41852ebb22c341c17fafba7e337a2af56f018dc1