Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Why is this change needed? Why would ro.build.id not change after OTA?
> Hi Amit
> 
> 
> We communicated with our OTA module colleagues.The property of ro.buil.id is defined in make/core/build_id.mk:export BUILD_ID=OPM2.171019.012
> and this property will change only when the Android version upgrade(like 7.0 to 8.1 , 8.1 to 9.0)
> However,When we push some modification through OTA and Android Version does not change, this property will not change. So,how should we upgarde the propery of ro.build.id under such circumstances ?

That build id looks like it should change with every OTA.

Jordan, can you please follow up on this when you get a chance? If it turns out that the right property to use is the one proposed in this CL, the change lgtm. Although we may get rid of ro.build.id check in that case.

Patch-set: 2
Reviewer: Gerrit User 1054648 <1054648@85c56323-6fa9-3386-8a01-6480fb634889>