Implement RcsProvider
This is a provider for internal use to access the RCS database tables.

This change is basically ag/5542367 with the crash fixed. In
RcsProvider, we now don't open the database immediately, instead wait
for it to be necessary to open the database.

Test: Added unit test, also tested manually by setting the flag to true
and seeing the database with the new table.

Bug: 109759350
Change-Id: I1538c873c116d3c27740745dea9e4761bdce0b09