Move RCS specific canonical address handling
This allows us to open up RCS provider and keep permissions the same for
the MMS provider.

Note that the way RCS message storage interacts with canonical-addresses
may change in the future. This commit fixes issues with the current
interaction with canonical-addresses.

Permissions around the provider will be addressed in b/124000636.

Test: CtsRcsTestCases
Bug: 122732718
Change-Id: I00e73dc50139599250dcf53fba23390a7cd6fe5d