Update patch set 3
Patch Set 3:

This still does not make sense, and as far as I am aware code has not changed. My guess
is that this is not used very often since it is obviously broken and has been for a long
time.

Patch-set: 3
Reviewer: Gerrit User 1004424 <1004424@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0