Mcc and mnc from xml in RRO may be integers
Mcc and mnc in an xml from overlays built in current android build
system are integers to the contrary of the apns-full-conf.xml
where they are strings. The telephony provider uses the same parsing
mechanism in both cases which makes the full apn list get
correct values whereas the provider misjudge what it should
insert in the database when there is an apns.xml in an overlay.

For the newly built android RROs mnc="01" becomes mnc="1" and no
match is found when the sim card is inserted.

To not risk the large APN list we only change the behavior for the
overlay handling, we use the locale list to get the correct mcc and
the carrier id table to get correct mnc.

Test: Manual
Test: adb shell content query --uri content://telephony/carriers
Bug: 175620622
Change-Id: I244ad0f76f360a0635343187e455b54c90ec45b4